created the exchange item

+++ home-helper-api/src/main/kotlin/home/helper/api/persistence/database/mysql/help_point/search/SearchHelpPointResult.kt
@@ -10,6 +10,6 @@ data class SearchHelpPointResult(
     var id: Long? = null,
     var userId: String? = null,
     var earnedDate: LocalDate? = null,
-    var earnedPoint: Long? = null,
+    var earnedPoint: Int? = null,
     var detailList: List<HelpItemResult>? = null,
 )
